+++ tools/perf/bench_v8/driver.js
@@ -1,0 +1,7 @@
+var success = true;
+function PrintResult(name, result) { print(name + ': ' + result); }
+function PrintError(name, error) { PrintResult(name, 'ERROR: ' + error); success = false; }
+function PrintScore(score) {
+    if (success) { print('----'); print('Score (version ' + BenchmarkSuite.version + '): ' + score); }
+}
+BenchmarkSuite.RunSuites({ NotifyResult: PrintResult, NotifyError: PrintError, NotifyScore: PrintScore });